An incomplete Game
This was coding as my self-learning iOS development. The idea and UI
were copied from an app store game. I’m opening this project for
everyone who is trying / learning objective-c. It is a typical MVC
project u can learn. However, it is welcome to any one who wants to
built this game perfect, give it modification make it better, and
strong it. So, any changes are welcome!